refactor(automation): decompose CIDriver into 4 SRP collaborators (#1179+#1289)
Extract four narrow collaborator modules from the CIDriver god-class (3,570 ln):

- hephaestus/automation/pr_discovery.py — PRDiscovery: viewer-login caching
  (#821) + PR enumeration (issue-driven, bot, failing)
- hephaestus/automation/ci_check_inspector.py — CICheckInspector + canonical
  FAILING_CHECK_CONCLUSIONS: required CI check queries via gh pr checks
- hephaestus/automation/ci_fix_orchestrator.py — CIFixOrchestrator: CI fix
  session management (codex + claude paths, push guards, prompt building)
- hephaestus/automation/post_merge_processor.py — PostMergeProcessor:
  drive-green /learn + /compact after PR merges

CIDriver becomes a thin facade of delegation stubs; each collaborator receives
only the narrow Callable[[], T] providers it needs (DIP, never self/CIDriver).
Lambda wrappers on all injected callables preserve patch.object(driver, "_X")
interception for the existing test suite (1,676 tests, 0 failures).

Orchestration methods (_attempt_ci_fixes, _retry_no_commit_once,
_run_ci_fix_session, _push_ci_fix) remain on CIDriver to maintain patch.object
test compatibility; their self._X() calls chain through delegation stubs.

Preservation note: _recheck_and_arm_after_fix (#1355), _resolve_dirty_pr
(#1347), _resolve_blocked_pr / _address_threads_once (#1356) kept with exact
behavior on CIDriver as required.

Closes #1357
