+++ tests/unit/automation/test_ci_driver_author_scope.py
@@ -105,9 +105,9 @@ class TestDefaultScopedDiscovery:
     def test_bot_pr_discovery_filters_to_viewer(self, viewer_driver: CIDriver) -> None:
         viewer_driver.options.include_all_authors = False
         with (
-            patch("hephaestus.automation.ci_driver.get_repo_info", return_value=("o", "r")),
+            patch("hephaestus.automation.pr_discovery.get_repo_info", return_value=("o", "r")),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 return_value=MagicMock(stdout=json.dumps(_MIXED_PULLS)),
             ),
         ):
@@ -116,9 +116,9 @@ class TestDefaultScopedDiscovery:
     def test_open_prs_remaining_filters_to_viewer(self, viewer_driver: CIDriver) -> None:
         viewer_driver.options.include_all_authors = False
         with (
-            patch("hephaestus.automation.ci_driver.get_repo_info", return_value=("o", "r")),
+            patch("hephaestus.automation.pr_discovery.get_repo_info", return_value=("o", "r")),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 return_value=MagicMock(stdout=json.dumps(_MIXED_PULLS)),
             ),
             patch.object(viewer_driver, "_pr_merge_state", return_value=("CLEAN", "MERGEABLE")),
@@ -133,9 +133,9 @@ class TestAllFlagScopedDiscovery:
     def test_bot_pr_discovery_returns_all_bots(self, driver: CIDriver) -> None:
         driver.options.include_all_authors = True
         with (
-            patch("hephaestus.automation.ci_driver.get_repo_info", return_value=("o", "r")),
+            patch("hephaestus.automation.pr_discovery.get_repo_info", return_value=("o", "r")),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 return_value=MagicMock(stdout=json.dumps(_MIXED_PULLS)),
             ),
         ):
@@ -144,9 +144,9 @@ class TestAllFlagScopedDiscovery:
     def test_open_prs_remaining_returns_all_prs(self, driver: CIDriver) -> None:
         driver.options.include_all_authors = True
         with (
-            patch("hephaestus.automation.ci_driver.get_repo_info", return_value=("o", "r")),
+            patch("hephaestus.automation.pr_discovery.get_repo_info", return_value=("o", "r")),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 return_value=MagicMock(stdout=json.dumps(_MIXED_PULLS)),
             ),
             patch.object(driver, "_pr_merge_state", return_value=("CLEAN", "MERGEABLE")),
@@ -173,7 +173,7 @@ class TestResolveViewerLogin:
     def test_resolve_caches_value(self, driver: CIDriver) -> None:
         driver._viewer_login = ""  # reset
         with patch(
-            "hephaestus.automation.ci_driver._gh_call", return_value=MagicMock(stdout="mvillmow\n")
+            "hephaestus.automation.pr_discovery._gh_call", return_value=MagicMock(stdout="mvillmow\n")
         ) as mock_gh:
             assert driver._resolve_viewer_login() == "mvillmow"
             assert driver._resolve_viewer_login() == "mvillmow"
@@ -182,7 +182,7 @@ class TestResolveViewerLogin:
     def test_resolve_failure_raises_runtimeerror(self, driver: CIDriver) -> None:
         driver._viewer_login = ""
         with patch(
-            "hephaestus.automation.ci_driver._gh_call",
+            "hephaestus.automation.pr_discovery._gh_call",
             side_effect=subprocess.CalledProcessError(1, ["gh"]),
         ):
             with pytest.raises(RuntimeError, match="Could not resolve viewer login"):
@@ -190,14 +190,14 @@ class TestResolveViewerLogin:
 
     def test_resolve_empty_stdout_raises(self, driver: CIDriver) -> None:
         driver._viewer_login = ""
-        with patch("hephaestus.automation.ci_driver._gh_call", return_value=MagicMock(stdout="")):
+        with patch("hephaestus.automation.pr_discovery._gh_call", return_value=MagicMock(stdout="")):
             with pytest.raises(RuntimeError, match="Could not resolve viewer login"):
                 driver._resolve_viewer_login()
 
     def test_resolve_gh_not_installed_raises(self, driver: CIDriver) -> None:
         driver._viewer_login = ""
         with patch(
-            "hephaestus.automation.ci_driver._gh_call",
+            "hephaestus.automation.pr_discovery._gh_call",
             side_effect=FileNotFoundError("gh not on PATH"),
         ):
             with pytest.raises(RuntimeError, match="Could not resolve viewer login"):
@@ -212,7 +212,7 @@ class TestResolveViewerLogin:
         """
         driver._viewer_login = ""
         with patch(
-            "hephaestus.automation.ci_driver._gh_call",
+            "hephaestus.automation.pr_discovery._gh_call",
             side_effect=GitHubUnavailableError("circuit breaker open"),
         ):
             with pytest.raises(RuntimeError, match="Could not resolve viewer login"):
@@ -230,9 +230,9 @@ class TestAllFlagSkipsViewerResolution:
                 "_resolve_viewer_login",
                 side_effect=RuntimeError("should not be called"),
             ),
-            patch("hephaestus.automation.ci_driver.get_repo_info", return_value=("o", "r")),
+            patch("hephaestus.automation.pr_discovery.get_repo_info", return_value=("o", "r")),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 return_value=MagicMock(stdout=json.dumps(_MIXED_PULLS)),
             ),
         ):
@@ -248,9 +248,9 @@ class TestAllFlagSkipsViewerResolution:
                 "_resolve_viewer_login",
                 side_effect=RuntimeError("should not be called"),
             ),
-            patch("hephaestus.automation.ci_driver.get_repo_info", return_value=("o", "r")),
+            patch("hephaestus.automation.pr_discovery.get_repo_info", return_value=("o", "r")),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 return_value=MagicMock(stdout=json.dumps(_MIXED_PULLS)),
             ),
             patch.object(driver, "_pr_merge_state", return_value=("CLEAN", "MERGEABLE")),
@@ -266,12 +266,12 @@ class TestMissingUserLogin:
         self, viewer_driver: CIDriver, caplog: pytest.LogCaptureFixture
     ) -> None:
         with (
-            patch("hephaestus.automation.ci_driver.get_repo_info", return_value=("o", "r")),
+            patch("hephaestus.automation.pr_discovery.get_repo_info", return_value=("o", "r")),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 return_value=MagicMock(stdout=json.dumps(_MISSING_LOGIN_PULLS)),
             ),
-            caplog.at_level(logging.WARNING, logger="hephaestus.automation.ci_driver"),
+            caplog.at_level(logging.WARNING, logger="hephaestus.automation.pr_discovery"),
         ):
             result = viewer_driver._list_open_prs_remaining()
 
@@ -285,12 +285,12 @@ class TestMissingUserLogin:
         self, viewer_driver: CIDriver, caplog: pytest.LogCaptureFixture
     ) -> None:
         with (
-            patch("hephaestus.automation.ci_driver.get_repo_info", return_value=("o", "r")),
+            patch("hephaestus.automation.pr_discovery.get_repo_info", return_value=("o", "r")),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 return_value=MagicMock(stdout=json.dumps(_MISSING_LOGIN_BOT_PULLS)),
             ),
-            caplog.at_level(logging.WARNING, logger="hephaestus.automation.ci_driver"),
+            caplog.at_level(logging.WARNING, logger="hephaestus.automation.pr_discovery"),
         ):
             result = viewer_driver._discover_bot_prs()
 
@@ -306,13 +306,13 @@ class TestMissingUserLogin:
         """Warning must NOT fire when --all is set (viewer filter bypassed entirely)."""
         driver.options.include_all_authors = True
         with (
-            patch("hephaestus.automation.ci_driver.get_repo_info", return_value=("o", "r")),
+            patch("hephaestus.automation.pr_discovery.get_repo_info", return_value=("o", "r")),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 return_value=MagicMock(stdout=json.dumps(_MISSING_LOGIN_PULLS)),
             ),
             patch.object(driver, "_pr_merge_state", return_value=("CLEAN", "MERGEABLE")),
-            caplog.at_level(logging.WARNING, logger="hephaestus.automation.ci_driver"),
+            caplog.at_level(logging.WARNING, logger="hephaestus.automation.pr_discovery"),
         ):
             result = driver._list_open_prs_remaining()
 
@@ -327,12 +327,12 @@ class TestMissingUserLogin:
         """Warning must NOT fire when --all is set (viewer filter bypassed entirely)."""
         driver.options.include_all_authors = True
         with (
-            patch("hephaestus.automation.ci_driver.get_repo_info", return_value=("o", "r")),
+            patch("hephaestus.automation.pr_discovery.get_repo_info", return_value=("o", "r")),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 return_value=MagicMock(stdout=json.dumps(_MISSING_LOGIN_BOT_PULLS)),
             ),
-            caplog.at_level(logging.WARNING, logger="hephaestus.automation.ci_driver"),
+            caplog.at_level(logging.WARNING, logger="hephaestus.automation.pr_discovery"),
         ):
             driver._discover_bot_prs()
