+++ tests/unit/automation/test_phase_agent_wiring.py
@@ -61,9 +61,16 @@ SELF_AGENT_PHASES: list[tuple[str, str, tuple[str, ...]]] = [
             "_review_phase.py",
         ),
     ),
-    # ci_driver owns Session 3 (AGENT_CI_DRIVER): its fix sessions and its
-    # post-green learnings run on a transcript independent of the implementer.
-    ("ci_driver.py", "AGENT_CI_DRIVER", ()),
+    # ci_driver owns Session 3 (AGENT_CI_DRIVER): its fix sessions (#1357
+    # extracted to ci_fix_orchestrator) and post-green learnings (#1357
+    # extracted to post_merge_processor) run on a transcript independent of
+    # the implementer. Companions hold the actual invoke_claude_with_session
+    # call sites after the SRP decomposition.
+    (
+        "ci_driver.py",
+        "AGENT_CI_DRIVER",
+        ("ci_fix_orchestrator.py", "post_merge_processor.py"),
+    ),
 ]
 
 
@@ -280,8 +287,8 @@ def test_per_iteration_reviewer_does_not_pin_foreign_agent(
 ADVISE_FIRST_STAGES_ADVISE_AGENT: list[tuple[str, tuple[str, ...]]] = [
     # Stage 1: the planner runs advise once before the plan loop.
     ("planner.py", ("planner_review_loop.py",)),
-    # Stage 3: the CI driver runs advise before the fix loop.
-    ("ci_driver.py", ()),
+    # Stage 3: the CI driver runs advise before the fix loop (impl in ci_fix_orchestrator).
+    ("ci_driver.py", ("ci_fix_orchestrator.py",)),
 ]
